feat: segment-count selector for planned solar system
Adds a 1/2/3 "Antal tag-flader" toggle to SolarConfigForm so users can
model east/west split roofs (#41). Extracts SliderField/TiltIllustration/
azimuthLabel into a shared module and a new RoofSegmentCard so the
per-segment fields aren't duplicated, sums total kWp across segments in
the sidebar and results header, and renders one direction arrow per
segment on the address map.
